// ps-crm/src/app/students/models/registrationTwo.model.ts
export interface RegistrationTwoModel {
  permit_issued: string;
  permit_expiry: string;
  license_num: string;
  license_type_o: string | boolean;
  license_type_a: string | boolean;
  license_type_b: string | boolean;
  license_type_c: string | boolean;
  endorsement_h: string | boolean;
  endorsement_n: string | boolean;
  endorsement_p: string | boolean;
  endorsement_s: string | boolean;
  endorsement_t: string | boolean;
  endorsement_x: string | boolean;
  endorsement_none: string | boolean;
  restriction_e: string | boolean;
  restriction_k: string | boolean;
  restriction_l: string | boolean;
  restriction_m: string | boolean;
  restriction_n: string | boolean;
  restriction_o: string | boolean;
  restriction_p: string | boolean;
  restriction_v: string | boolean;
  restriction_x: string | boolean;
  restriction_z: string | boolean;
  restriction_none: string | boolean;
}